more readable component value implementation
* get rid of convert_case dependency for now
